DefaultUM7 class almoast ready for debugging (without parsedatabatch and catchSample)

+++ src/main/java/pl/agilevision/hardware/um7/impl/UM7Constants.java
@@ -9,13 +9,13 @@ public interface UM7Constants {
    * Commands
    */
   interface Commands {
-    int GET_FW_REVISION       = 0xAA; //# (170)
-    int FLASH_COMMIT          = 0xAB; //# (171)
-    int RESET_TO_FACTORY      = 0xAC; //# (172)
-    int ZERO_GYROS            = 0xAD; //# (173)
-    int SET_HOME_POSITION     = 0xAE; //# (174)
-    int SET_MAG_REFERENCE     = 0xB0; //# (176)
-    int RESET_EKF             = 0xB3; //# (179)
+    byte GET_FW_REVISION       = (byte)0xAA; //# (170)
+    byte FLASH_COMMIT          = (byte)0xAB; //# (171)
+    byte RESET_TO_FACTORY      = (byte)0xAC; //# (172)
+    byte ZERO_GYROS            = (byte)0xAD; //# (173)
+    byte SET_HOME_POSITION     = (byte)0xAE; //# (174)
+    byte SET_MAG_REFERENCE     = (byte)0xB0; //# (176)
+    byte RESET_EKF             = (byte)0xB3; //# (179)
 
   }
 
@@ -23,23 +23,23 @@ public interface UM7Constants {
    * Registers
    */
   interface Registers {
-    int DREG_HEALTH           = 0x55;
-    int DREG_GYRO_RAW_XY      = 0x56;
-    int DREG_GYRO_PROC_X      = 0x61;
-    int DREG_ACCEL_PROC_X     = 0x65;
-    int DREG_EULER_PHI_THETA  = 0x70;
-    int DREG_GYRO_BIAS_X      = 0x89;
+    byte DREG_HEALTH           = (byte)0x55;
+    byte DREG_GYRO_RAW_XY      = (byte)0x56;
+    byte DREG_GYRO_PROC_X      = (byte)0x61;
+    byte DREG_ACCEL_PROC_X     = (byte)0x65;
+    byte DREG_EULER_PHI_THETA  = (byte)0x70;
+    byte DREG_GYRO_BIAS_X      = (byte)0x89;
 
-    int CREG_COM_SETTINGS     = 0x00;
-    int CREG_GYRO_TRIM_X      = 0x0C;
-    int CREG_MAG_CAL1_1       = 0x0F;
-    int CREG_MAG_BIAS_X       = 0x18;
+    byte CREG_COM_SETTINGS     = (byte)0x00;
+    byte CREG_GYRO_TRIM_X      = (byte)0x0C;
+    byte CREG_MAG_CAL1_1       = (byte)0x0F;
+    byte CREG_MAG_BIAS_X       = (byte)0x18;
 
-    int REG_HIDDEN            = 0xF000;
-    int H_CREG_GYRO_ALIGN1_1  = REG_HIDDEN | 0x31;
-    int H_CREG_ACCEL_ALIGN1_1 = REG_HIDDEN | 0x52;
-    int H_CREG_MAG_ALIGN1_1   = REG_HIDDEN | 0x73;
-    int H_CREG_MAG_REF        = REG_HIDDEN | 0x7C;
+    short REG_HIDDEN            = (short) 0xF000;
+    int H_CREG_GYRO_ALIGN1_1  = REG_HIDDEN | (byte)0x31;
+    int H_CREG_ACCEL_ALIGN1_1 = REG_HIDDEN | (byte)0x52;
+    int H_CREG_MAG_ALIGN1_1   = REG_HIDDEN | (byte)0x73;
+    int H_CREG_MAG_REF        = REG_HIDDEN | (byte)0x7C;
   }
 
   /**
